Separate ECMAScript into its own directory
