+++ lib/ECMAScript/ECMAScriptLexer.interp
