+++ lib/ECMAScript/ECMAScriptLexer.js
@@ -1,6 +1,6 @@
 // Generated from grammars/ECMAScript.g4 by ANTLR 4.7.2
 // jshint ignore: start
-var antlr4 = require('../lib/antlr4/index');
+var antlr4 = require('../../lib/antlr4/index');
