+++ lib/ECMAScript/ECMAScriptVisitor.js
@@ -1,6 +1,6 @@
 // Generated from grammars/ECMAScript.g4 by ANTLR 4.7.2
 // jshint ignore: start
-var antlr4 = require('../lib/antlr4/index');
+var antlr4 = require('../../lib/antlr4/index');
 
 // This class defines a complete generic visitor for a parse tree produced by ECMAScriptParser.
