+++ makefile
@@ -1,2 +1,2 @@
-bundle :
-	browserify index.js -o build/bundle.js
+bundle_ecma :
+	browserify ecma_index.js -o build/bundle.js
